+++ twilio/rest/verify/v2/service/entity/factor.py
@@ -35,58 +35,6 @@ class FactorList(ListResource):
         self._solution = {'service_sid': service_sid, 'identity': identity, }
         self._uri = '/Services/{service_sid}/Entities/{identity}/Factors'.format(**self._solution)
 
-    def create(self, friendly_name, factor_type, binding_alg=values.unset,
-               binding_public_key=values.unset, config_app_id=values.unset,
-               config_notification_platform=values.unset,
-               config_notification_token=values.unset,
-               config_sdk_version=values.unset, binding_secret=values.unset,
-               config_time_step=values.unset, config_skew=values.unset,
-               config_code_length=values.unset, config_alg=values.unset):
-        """
-        Create the FactorInstance
-
-        :param unicode friendly_name: The friendly name of this Factor
-        :param FactorInstance.FactorTypes factor_type: The Type of this Factor
-        :param unicode binding_alg: The algorithm used when `factor_type` is `push`
-        :param unicode binding_public_key: The public key encoded in Base64
-        :param unicode config_app_id: The ID that uniquely identifies your app in the Google or Apple store
-        :param FactorInstance.NotificationPlatforms config_notification_platform: The transport technology used to generate the Notification Token
-        :param unicode config_notification_token: For APN, the device token. For FCM the registration token
-        :param unicode config_sdk_version: The Verify Push SDK version used to configure the factor
-        :param unicode binding_secret: The shared secret in Base32
-        :param unicode config_time_step: How often, in seconds, are TOTP codes generated
-        :param unicode config_skew: The number of past and future time-steps valid at a given time
-        :param unicode config_code_length: Number of digits for generated TOTP codes
-        :param FactorInstance.TotpAlgorithms config_alg: The algorithm used to derive the TOTP codes
-
-        :returns: The created FactorInstance
-        :rtype: twilio.rest.verify.v2.service.entity.factor.FactorInstance
-        """
-        data = values.of({
-            'FriendlyName': friendly_name,
-            'FactorType': factor_type,
-            'Binding.Alg': binding_alg,
-            'Binding.PublicKey': binding_public_key,
-            'Config.AppId': config_app_id,
-            'Config.NotificationPlatform': config_notification_platform,
-            'Config.NotificationToken': config_notification_token,
-            'Config.SdkVersion': config_sdk_version,
-            'Binding.Secret': binding_secret,
-            'Config.TimeStep': config_time_step,
-            'Config.Skew': config_skew,
-            'Config.CodeLength': config_code_length,
-            'Config.Alg': config_alg,
-        })
-
-        payload = self._version.create(method='POST', uri=self._uri, data=data, )
-
-        return FactorInstance(
-            self._version,
-            payload,
-            service_sid=self._solution['service_sid'],
-            identity=self._solution['identity'],
-        )
-
     def stream(self, limit=None, page_size=None):
         """
         Streams FactorInstance records from the API as a generator stream.